+++ bpf/insn/samples/strlen.c
@@ -1,7 +1,7 @@
 /**
- * int bpf_strnstr(const char *s1__ign, const char *s2__ign, size_t len)
+ * int bpf_strlen(const char *s__ign);
  *
- * - https://docs.ebpf.io/linux/kfuncs/bpf_strnstr/
+ * - https://docs.ebpf.io/linux/kfuncs/bpf_strlen/
  */
 #include <stdio.h>
 #include "bpf_insn_samples.h"
@@ -10,12 +10,12 @@
 
 static struct bpf_insn insns_buf[1024];
 
-BPF_INSN_SAMPLE_FUNC_PROTO(strnstr)
+BPF_INSN_SAMPLE_FUNC_PROTO(strlen)
 {
-	int btf_id = btf_has_kfunc(NULL, "bpf_strnstr", true);
+	int btf_id = btf_has_kfunc(NULL, "bpf_strlen", true);
 
 	if (btf_id <= 0) {
-		fprintf(stderr, "ERROR: not found strnstr kfunc.\n");
+		fprintf(stderr, "ERROR: not found strlen kfunc.\n");
 		*cnt = 0;
 		return NULL;
 	}
@@ -23,8 +23,6 @@ BPF_INSN_SAMPLE_FUNC_PROTO(strnstr)
 	struct bpf_insn *insn = insns_buf;
 
 	*insn++ = BPF_MOV64_IMM(BPF_REG_1, 0);
-	*insn++ = BPF_MOV64_IMM(BPF_REG_2, 0);
-	*insn++ = BPF_MOV64_IMM(BPF_REG_3, 0); // len must be zero here
 	*insn++ = BPF_CALL_KFUNC(0, btf_id);
 	*insn++ = BPF_MOV64_IMM(BPF_REG_0, 0);
 	*insn++ = BPF_EXIT_INSN();