+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to
 ## Unreleased
 #### Breaking Changes
 #### Added
+- bpf: kfunc: test `bpf_strlen()` and `bpf_strnlen()`;
 - bpftrace/samples: add string subdir;
 - rpm: add spec subdir;
 - add `__VA_OPT__.cpp` test;